+++ src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioTest.java
@@ -2,7 +2,7 @@ package ar.edu.unlam.tallerweb1.servicios;
 
 public interface ServicioTest {
 
-	Boolean realizarTest(Integer fiebre,Boolean olfato,Boolean gusto,Boolean tos , Boolean respiracion);
+	Boolean realizarTest(Float fiebre,Boolean olfato,Boolean gusto,Boolean tos , Boolean respiracion);
